splitted archiveEserviceInternal in two functions for better execution management
